% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt_lm.R
\name{chatgpt_lm}
\alias{chatgpt_lm}
\title{Fit a linear regression model using ChatGPT}
\usage{
chatgpt_lm(df, predict_col)
}
\arguments{
\item{df}{a data.frame}

\item{predict_col}{a string with the name of a variable in \code{df} to predict}
}
\description{
Fit a linear regression model using ChatGPT
}
