% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt_lm.R
\name{chatgpt_lm_prompt}
\alias{chatgpt_lm_prompt}
\title{Generate a prompt to get an lm() call from ChatGPT}
\usage{
chatgpt_lm_prompt(df, predict_col)
}
\description{
Generate a prompt to get an lm() call from ChatGPT
}
