% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt_lm.R
\name{chatgpt_lm_prompt}
\alias{chatgpt_lm_prompt}
\title{Generate a prompt to get an lm() call from ChatGPT}
\usage{
chatgpt_lm_prompt(df, predict_col)
}
\arguments{
\item{df}{a data.frame}

\item{predict_col}{a string with the name of a variable in \code{df} to predict}
}
\description{
Generate a prompt to get an lm() call from ChatGPT
}
